% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceedance.R
\name{exceedance_flow}
\alias{exceedance_flow}
\title{Exceedance Flow}
\usage{
exceedance_flow(flow, prob)
}
\arguments{
\item{flow}{Flow time series}

\item{prob}{Exceedance probability}
}
\description{
Calculate flow value that meets a given exceedance probability
}
\examples{
exceedance_flow(runif(30, 0, 10000), 0.5)

}
