% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceedance.R
\name{exceedance_prob}
\alias{exceedance_prob}
\title{Exceedance Probability}
\usage{
exceedance_prob(flow)
}
\arguments{
\item{flow}{Flow time series}
}
\description{
Calculate exceedance probability for a flow time series
}
\examples{
exceedance_prob(runif(30, 0, 10000))

}
