% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wy_type_table.R
\name{get_wy_type_table}
\alias{get_wy_type_table}
\title{Water year classifications}
\usage{
get_wy_type_table()
}
\description{
Get table of Water Year Hydrologic Classification Indices from \href{https://cdec.water.ca.gov/reportapp/javareports?name=WSIHIST}{CDEC}.
}
\seealso{
\link{get_water_year_type}, \link{water_year_type}
}
