% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egnyte.R
\name{split_egnyte_path}
\alias{split_egnyte_path}
\title{Split Egnyte path into parts}
\usage{
split_egnyte_path()
}
\description{
Split a file path copied from Windows File Explorer into parts for pasting into call to file.path().
Needs to be read from clipboard to escape the path separators.
}
