% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceedance.R
\name{exceedance_flow}
\alias{exceedance_flow}
\title{Exceedance Flow}
\usage{
exceedance_flow(flow, probs)
}
\arguments{
\item{flow}{Flow time series}

\item{probs}{Exceedance probabilities}
}
\description{
Calculate flow value that meet a vector of exceedance probabilities
}
\examples{
exceedance_flow(runif(30, 0, 10000), seq(0.1, 0.9, 0.1))

}
